Some more refactoring
Some more refactoring
