+++ sonar-project.properties
@@ -7,5 +7,5 @@ sonar.sources=src
 sonar.projectName=TourLive React Frontend
 sonar.sourceEncoding=UTF-8
 sonar.javascript.file.suffixes=.js,.jsx
-sonar.exclusions=node_modules/**, src/registerServiceWorker.js, src/index.js
+sonar.exclusions=node_modules/**, src/registerServiceWorker.js, src/index.js, src/util/countries.js
 sonar.javascript.lcov.reportPath=coverage/lcov.info
